console/console-count-logging.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test console log are present</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
    promise_test(async test => {
        await test_driver.bidi.log.entry_added.subscribe();

        const log_entries = [];
        // Collect all the `count` log entries' texts into `log_entries`.
        test_driver.bidi.log.entry_added.on(event => {
                if (event.method === 'count') {
                    log_entries.push(event.text);
                }
            }
        );

        console.count();
        console.count(undefined);
        console.count("default");
        console.count({
            toString() {
                return "default"
            }
        });

        await test.step_wait(() => log_entries.length >= 4,
            "Wait for all the log entries to be received");

        assert_array_equals(log_entries, [
            "default: 1",
            "default: 2",
            "default: 3",
            "default: 4"]);
    }, "Console count method default parameter should work");
</script>
